move InferenceModel to a different file
