+++ src/anomalib/deploy/export.py
@@ -15,11 +15,12 @@ from typing import TYPE_CHECKING, Any
 import numpy as np
 import torch
 from torch import nn
-from torchvision.transforms.v2 import CenterCrop, Compose, Resize, Transform
+from torchvision.transforms.v2 import Transform
 
 from anomalib import TaskType
-from anomalib.data.transforms import ExportableCenterCrop
 from anomalib.utils.exceptions import try_import
+
+from .inference_model import InferenceModel
 
 if TYPE_CHECKING:
     from torch.types import Number
@@ -43,59 +44,6 @@ class ExportType(str, Enum):
     ONNX = "onnx"
     OPENVINO = "openvino"
     TORCH = "torch"
-
-
-class InferenceModel(nn.Module):
-    """Inference model for export.
-
-    The InferenceModel is used to wrap the model and transform for exporting to torch and ONNX/OpenVINO.
-
-    Args:
-        model (nn.Module): Model to export.
-        transform (Transform): Input transform for the model.
-        disable_antialias (bool, optional): Disable antialiasing in the Resize transforms of the given transform. This
-            is needed for ONNX/OpenVINO export, as antialiasing is not supported in the ONNX opset.
-    """
-
-    def __init__(self, model: nn.Module, transform: Transform, disable_antialias: bool = False) -> None:
-        super().__init__()
-        self.model = model
-        self.transform = transform
-        self.convert_center_crop()
-        if disable_antialias:
-            self.disable_antialias()
-
-    def forward(self, batch: torch.Tensor) -> torch.Tensor | tuple[torch.Tensor, torch.Tensor]:
-        """Transform the input batch and pass it through the model."""
-        batch = self.transform(batch)
-        return self.model(batch)
-
-    def disable_antialias(self) -> None:
-        """Disable antialiasing in the Resize transforms of the given transform.
-
-        This is needed for ONNX/OpenVINO export, as antialiasing is not supported in the ONNX opset.
-        """
-        if isinstance(self.transform, Resize):
-            self.transform.antialias = False
-        if isinstance(self.transform, Compose):
-            for transform in self.transform.transforms:
-                if isinstance(transform, Resize):
-                    transform.antialias = False
-
-    def convert_center_crop(self) -> None:
-        """Convert CenterCrop to ExportableCenterCrop for ONNX export.
-
-        The original CenterCrop transform is not supported in ONNX export. This method replaces the CenterCrop to
-        ExportableCenterCrop, which is supported in ONNX export. For more details, see the implementation of
-        ExportableCenterCrop.
-        """
-        if isinstance(self.transform, CenterCrop):
-            self.transform = ExportableCenterCrop(size=self.transform.size)
-        elif isinstance(self.transform, Compose):
-            transforms = self.transform.transforms
-            for index in range(len(transforms)):
-                if isinstance(transforms[index], CenterCrop):
-                    transforms[index] = ExportableCenterCrop(size=transforms[index].size)
 
 
 class ExportMixin:
